rename: 도메인들 도메인 패키지로 이동
