+++ src/main/java/racingcar/model/domain/AllRacingCars.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.domain;
 
 import java.util.List;
