+++ src/main/java/racingcar/model/domain/RacingCar.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.domain;
 
 public class RacingCar {
